requirements.txt
mediapipe==0.10.21
numpy==1.26.4
opencv-python==4.11.0.86
paho-mqtt==2.1.0
PyYAML==6.0.2
requests==2.32.3
tensorflow==2.19.0
flask==3.1.0
